add lua-recastnavigator

+++ lualib-src/lua-recastnavigation/recastnavigation/Detour/Include/DetourAlloc.h
@@ -19,8 +19,6 @@
 #ifndef DETOURALLOCATOR_H
 #define DETOURALLOCATOR_H
 
-#include <stddef.h>
-
 /// Provides hint values to the memory allocator on how long the
 /// memory is expected to be used.
 enum dtAllocHint
@@ -34,7 +32,7 @@ enum dtAllocHint
 //  @param[in]		rcAllocHint	A hint to the allocator on how long the memory is expected to be in use.
 //  @return A pointer to the beginning of the allocated memory block, or null if the allocation failed.
 ///  @see dtAllocSetCustom
-typedef void* (dtAllocFunc)(size_t size, dtAllocHint hint);
+typedef void* (dtAllocFunc)(int size, dtAllocHint hint);
 
 /// A memory deallocation function.
 ///  @param[in]		ptr		A pointer to a memory block previously allocated using #dtAllocFunc.
@@ -51,7 +49,7 @@ void dtAllocSetCustom(dtAllocFunc *allocFunc, dtFreeFunc *freeFunc);
 ///  @param[in]		hint	A hint to the allocator on how long the memory is expected to be in use.
 ///  @return A pointer to the beginning of the allocated memory block, or null if the allocation failed.
 /// @see dtFree
-void* dtAlloc(size_t size, dtAllocHint hint);
+void* dtAlloc(int size, dtAllocHint hint);
 
 /// Deallocates a memory block.
 ///  @param[in]		ptr		A pointer to a memory block previously allocated using #dtAlloc.
